Refactor test functions (#565)
Splitting out some work from #557

---------
