+++ mlos_bench/mlos_bench/tests/environments/local/__init__.py
@@ -6,12 +6,9 @@
 Tests for mlos_bench.environments.local.
 Used to make mypy happy about multiple conftest.py modules.
 """
-from datetime import datetime
-from typing import Any, Dict, List, Tuple
 
-import pytest
+from typing import Any, Dict, List
 
-from mlos_bench.environments.base_environment import Environment
 from mlos_bench.environments.composite_env import CompositeEnv
 from mlos_bench.environments.local.local_env import LocalEnv
 from mlos_bench.services.config_persistence import ConfigPersistenceService
@@ -79,56 +76,3 @@ def create_composite_local_env(tunable_groups: TunableGroups,
         global_config=global_config,
         service=LocalExecService(parent=ConfigPersistenceService()),
     )
-
-
-def check_env_success(env: Environment,
-                      tunable_groups: TunableGroups,
-                      expected_results: Dict[str, float],
-                      expected_telemetry: List[Tuple[datetime, str, Any]]) -> None:
-    """
-    Set up a local environment and run a test experiment there.
-
-    Parameters
-    ----------
-    tunable_groups : TunableGroups
-        Tunable parameters (usually come from a fixture).
-    env : Environment
-        An environment to query for the results.
-    expected_results : Dict[str, float]
-        Expected results of the benchmark.
-    expected_telemetry : List[Tuple[datetime, str, Any]]
-        Expected telemetry data of the benchmark.
-    """
-    with env as env_context:
-
-        assert env_context.setup(tunable_groups)
-
-        (status, data) = env_context.run()
-        assert status.is_succeeded()
-        assert data == pytest.approx(expected_results, nan_ok=True)
-
-        (status, telemetry) = env_context.status()
-        assert status.is_good()
-        assert telemetry == pytest.approx(expected_telemetry, nan_ok=True)
-
-
-def check_env_fail_telemetry(env: Environment, tunable_groups: TunableGroups) -> None:
-    """
-    Set up a local environment and run a test experiment there;
-    Make sure the environment `.status()` call fails.
-
-    Parameters
-    ----------
-    tunable_groups : TunableGroups
-        Tunable parameters (usually come from a fixture).
-    env : Environment
-        An environment to query for the results.
-    """
-    with env as env_context:
-
-        assert env_context.setup(tunable_groups)
-        (status, _data) = env_context.run()
-        assert status.is_succeeded()
-
-        with pytest.raises(ValueError):
-            env_context.status()
