+++ mlos_bench/mlos_bench/tests/environments/local/local_env_stdout_test.py
@@ -7,7 +7,8 @@ Unit tests for extracting data from LocalEnv stdout.
 """
 
 from mlos_bench.tunables.tunable_groups import TunableGroups
-from mlos_bench.tests.environments.local import create_local_env, check_env_success
+from mlos_bench.tests.environments import check_env_success
+from mlos_bench.tests.environments.local import create_local_env
 
 
 def test_local_env_stdout(tunable_groups: TunableGroups) -> None:
