+++ mlos_bench/mlos_bench/tests/environments/local/local_env_telemetry_test.py
@@ -8,9 +8,8 @@ Unit tests for telemetry and status of LocalEnv benchmark environment.
 from datetime import datetime, timedelta
 
 from mlos_bench.tunables.tunable_groups import TunableGroups
-from mlos_bench.tests.environments.local import (
-    create_local_env, check_env_success, check_env_fail_telemetry
-)
+from mlos_bench.tests.environments import check_env_success, check_env_fail_telemetry
+from mlos_bench.tests.environments.local import create_local_env
 
 
 def test_local_env_telemetry(tunable_groups: TunableGroups) -> None:
